refactor(base): refactored base folder
