+++ src/base/config/git.ts
@@ -1,5 +1,5 @@
-import { Config } from './config';
-import { ProjectTypes } from '../types/project';
+import { Config } from '../config';
+import { ProjectTypes } from '../../types/project';
 
 /**
  * Base class for implementing all relevant Git configuration.
