+++ src/base/index.ts
@@ -1,15 +1,15 @@
 export * from './config';
-export * from './commitlint';
-export * from './devcontainer';
-export * from './eslint';
-export * from './git';
-export * from './github';
-export * from './husky';
-export * from './jest';
-export * from './npm';
-export * from './options';
+export * from './config/commitlint';
+export * from './config/devcontainer';
+export * from './config/eslint';
+export * from './config/git';
+export * from './config/github';
+export * from './config/husky';
+export * from './config/jest';
+export * from './config/npm';
+export * from './config/options';
 export * from './project';
-export * from './prettier';
+export * from './config/prettier';
 export * from './typescript';
-export * from './vscode';
-export * from './samplecode';
+export * from './config/vscode';
+export * from './config/samplecode';
